revert: restore tee-writer/trace-collector schema splits (400-LOC rule)
Project eslint rule caps files at 400 LOC. Merged tee-writer + schema (537)
and trace-collector + schema (600) tripped the limit. Keep the facilitator
merge (380 LOC, under limit). Net: -1 file from the suite instead of -3.
